+++ CHANGELOG.md
@@ -11,9 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Adds elastica raw result in serializer context while denormalizing results
+
 ### Deprecated
 
-* deprecates `Index::getBuilder()` in favor of `Client::getBuilder()`
+- Deprecates `Index::getBuilder()` in favor of `Client::getBuilder()`
 
 ## [1.2.0] - 2021-02-16